requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
tensorflow==2.12.0
python-dotenv==1.0.0
flask==2.3.2
